+++ include/autd3/special_data.hpp
@@ -3,7 +3,7 @@
 // Created Date: 07/11/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 07/11/2022
+// Last Modified: 08/11/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -13,6 +13,7 @@
 
 #include <autd3/core/amplitudes.hpp>
 #include <autd3/core/clear.hpp>
+#include <autd3/core/delay.hpp>
 #include <autd3/core/silencer_config.hpp>
 #include <autd3/core/synchronize.hpp>
 
@@ -68,5 +69,23 @@ class Synchronize : public SpecialData<core::NullHeader, core::Synchronize> {
   [[nodiscard]] bool check_trials_override() const override { return true; }
   [[nodiscard]] size_t check_trials() const override { return 200; }
 };
+
+class ModDelayConfig : public SpecialData<core::NullHeader, core::ModDelayConfig> {
+ public:
+  ModDelayConfig() : SpecialData<header_t, body_t>(core::NullHeader{}, core::ModDelayConfig{}) {}
+
+  [[nodiscard]] bool check_trials_override() const override { return false; }
+  [[nodiscard]] size_t check_trials() const override { return 0; }
+};
+
+inline autd3::Stop stop() { return autd3::Stop{}; }
+
+inline autd3::UpdateFlag update_flag() { return autd3::UpdateFlag{}; }
+
+inline autd3::Clear clear() { return autd3::Clear{}; }
+
+inline autd3::Synchronize synchronize() { return autd3::Synchronize{}; }
+
+inline autd3::ModDelayConfig mod_delay_config() { return autd3::ModDelayConfig{}; }
 
 }  // namespace autd3