+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 - Fix [#37](https://github.com/shinolab/autd3/issues/37)
 - Fix [#42](https://github.com/shinolab/autd3/issues/42)
 - Change phase unit to radian
-- Add `autd3::Async` to send data asynchronously
-- `Controller::synchronize`, `Controller::update_flag`, `Controller::clear`, and `Controller::stop` are now deprecated
-  - Send `autd3::Synchronize`, `autd3::UpdateFlag`, `autd3::Clear`, and `autd3::Stop` instead
+- Add stream operator to `Controller`
+- Add `Controller::send_async` and `autd3::async` to send data asynchronously
+- `Controller::synchronize()`, `Controller::update_flag()`, `Controller::clear()`, and `Controller::stop()` functions are now deprecated
+  - Send `autd3::synchronize`, `autd3::update_flag`, `autd3::clear`, and `autd3::stop` instead
 
 # 2.5.1